Extract trait options validating
